Quickly sketch out reverse auth proxy for NS
It seems to work: logged in allows access to Nightscout target, logging out
reverts back to Drywall display.

Thanks to @xitation for the idea in gitter.

One small problem, the target Nightscout sets it's own CSRF cookies, which
clobber Drywall's CSRF cookies, and makes node unhappy, to the point of dying.

Otherwise, all the components themselves look pretty good.  If that bug can be
fixed, this might make or an easy way to bolt on user auth in front of
something like Nightscout.